Create a module for deprecated tools
